+++ packages/picasso-lab/src/OverviewBlock/OverviewBlock.tsx
@@ -25,17 +25,11 @@ import { useOverviewBlockGroupContext } from '../OverviewBlockGroup/OverviewBloc
 import OverviewBlockGroup from '../OverviewBlockGroup'
 import OverviewBlockRow from '../OverviewBlockRow'
 
-type Variant =
-  | 'value-red'
-  | 'value-green'
-  | 'value-blue'
-  | 'value-yellow'
-  | 'label-red'
-  | 'label-green'
-  | 'label-blue'
-  | 'label-yellow'
+type VariantColorType = Extract<ColorType, 'red' | 'green' | 'yellow'>
 
-type ColorSettings = {
+type Variant = `value-${VariantColorType}` | `label-${VariantColorType}`
+
+type ColorSchema = {
   value: ColorType
   label: ColorType
 }
@@ -80,14 +74,14 @@ export const OverviewBlock: OverridableComponent<Props> & StaticProps =
     const classes = useStyles()
     const { align, blockWidth } = useOverviewBlockGroupContext()
 
-    const color: ColorSettings = {
+    const color: ColorSchema = {
       value: 'black',
       label: 'dark-grey'
     }
 
     if (variant) {
       const [partName, colorName] = variant.split('-') as [
-        keyof ColorSettings,
+        keyof ColorSchema,
         ColorType
       ]