+++ packages/picasso/src/Typography/story/index.jsx
@@ -12,23 +12,23 @@ page
 
 page
   .createChapter()
-  .addExample('Typography/story/Default.example.tsx', 'Normal text')
-  .addExample('Typography/story/Headings.example.tsx', 'Headings')
+  .addExample('Typography/story/Default.example.tsx', 'Normal text') // picasso-skip-visuals
+  .addExample('Typography/story/Headings.example.tsx', 'Headings') // picasso-skip-visuals
   .addExample('Typography/story/Types.example.tsx', {
     title: 'Types',
     description:
       'Long-form text uses a 1.5 ratio to calculate line-height values.'
-  })
+  }) // picasso-skip-visuals
   .addExample('Typography/story/InheritSize.example.tsx', {
     title: 'Inherit size',
     description: `
       In some cases, you may want to use the font size inherited
       from the container where the Topography is nested.
     `
-  })
-  .addExample('Typography/story/Alignment.example.tsx', 'Alignment')
-  .addExample('Typography/story/Weights.example.tsx', 'Weights')
-  .addExample('Typography/story/Colors.example.tsx', 'Colors')
+  }) // picasso-skip-visuals
+  .addExample('Typography/story/Alignment.example.tsx', 'Alignment') // picasso-skip-visuals
+  .addExample('Typography/story/Weights.example.tsx', 'Weights') // picasso-skip-visuals
+  .addExample('Typography/story/Colors.example.tsx', 'Colors') // picasso-skip-visuals
   .addExample('Typography/story/As.example.tsx', {
     title: 'Augment as another HTML element',
     description: `
@@ -36,6 +36,6 @@ page
       HTML tag used for Typography. In this cases you can 
       use 'as' property.
     `
-  })
-  .addExample('Typography/story/Decoration.example.tsx', 'Decoration')
-  .addExample('Typography/story/LineThrough.example.tsx', 'Line through')
+  }) // picasso-skip-visuals
+  .addExample('Typography/story/Decoration.example.tsx', 'Decoration') // picasso-skip-visuals
+  .addExample('Typography/story/LineThrough.example.tsx', 'Line through') // picasso-skip-visuals